perf: optimize CV page performance with lazy loading and code splitting
- Implement lazy loading for react-pdf dependencies
  - Only load react-pdf when custom CV data is used (not for default data)
  - Reduce initial bundle size from ~591 kB to ~11 kB for CV page
  - Buffer polyfill only loaded when react-pdf is needed
- Add code splitting configuration in Vite
  - Separate chunks for react-pdf (1.5 MB, on-demand)
  - Separate chunks for docx generation (355 kB, on-demand)
  - Separate chunks for CV editor (72 kB, edit mode only)
- Create CvDownloadButtonsCustom component
  - Extracted heavy dependencies into separate lazy-loaded component
  - Default data uses static file downloads (no heavy dependencies)
  - Custom data triggers lazy loading of react-pdf and docx libraries
- Lazy load CVEditor component
  - Only loaded when edit mode is activated
  - Reduces initial bundle size significantly
- Performance improvements
  - ~98% bundle size reduction for normal users
  - Eliminates 24+ second load times for CV page
  - Better caching strategy through separate chunks
